+++ src/Umbraco.Core/Umbraco.Core.csproj
@@ -132,13 +132,8 @@
     <Compile Include="Cache\RepositoryCachePolicyBase.cs" />
     <Compile Include="Cache\SingleItemsOnlyRepositoryCachePolicy.cs" />
     <Compile Include="Compose\AuditEventsComposer.cs" />
-    <Compile Include="Composing\ComponentComposer.cs" />
-    <Compile Include="Composing\ComposeAfterAttribute.cs" />
-    <Compile Include="Composing\ComposeBeforeAttribute.cs" />
-    <Compile Include="Composing\Composers.cs" />
     <Compile Include="Composing\RegisterFactory.cs" />
     <Compile Include="CompositionExtensions.cs" />
-    <Compile Include="Composing\IUserComposer.cs" />
     <Compile Include="Compose\ManifestWatcherComposer.cs" />
     <Compile Include="Compose\RelateOnCopyComposer.cs" />
     <Compile Include="Compose\RelateOnTrashComposer.cs" />
